I've refactored and simplified the application.
This change refactors the navigation in `Layout.js` to use a single array of links and conditional rendering, removes unused navigation links, extracts the dark mode logic into a custom `useDarkMode` hook, and simplifies the `CatchAll` component in `AppRoutes.js`.

Note: The changes in this commit have not been tested due to issues with the build and test environment. I recommend that you investigate the timeouts in `npm test` and `npm run build`.
